+++ tests/testthat/test-aes-calculated.R
@@ -99,3 +99,28 @@ test_that("A deprecated warning is issued when stat(var) or ..var.. is used", {
   p2 <- ggplot(NULL, aes(..bar..))
   expect_snapshot_warning(b2 <- ggplot_build(p2))
 })
+
+test_that("functions can be masked", {
+
+  foo <- function(x) x + 10
+  bar <- function(x) x * 10
+
+  data <- data.frame(val = 10)
+  mapping <- aes(x = val, y = foo(20))
+
+  evaled  <- eval_aesthetics(mapping, data = data, mask = list())
+  expect_equal(evaled, list(x = 10, y = 30))
+
+  evaled <- eval_aesthetics(mapping, data = data, mask = list(foo = bar))
+  expect_equal(evaled, list(x = 10, y = 200))
+
+  # Test namespace-prefixed evaluation (#6104)
+  mapping <- aes(x = val, y = ggplot2::stage(10, 20, 30))
+  evaled <- eval_aesthetics(mapping, data = data, mask = list())
+  expect_equal(evaled, list(x = 10, y = 10))
+  evaled <- eval_aesthetics(mapping, data = data, mask = list(stage = stage_calculated))
+  expect_equal(evaled, list(x = 10, y = 20))
+  evaled <- eval_aesthetics(mapping, data = data, mask = list(stage = stage_scaled))
+  expect_equal(evaled, list(x = 10, y = 30))
+
+})